Create a new shared component and a wrapper in app/web
