+++ apps/web/src/accessibility/roving/types.ts
@@ -6,4 +6,4 @@ SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Com
 Please see LICENSE files in the repository root for full details.
 */
 
-export type FocusHandler = () => void;
+export type { FocusHandler } from "@element-hq/web-shared-components";
