+++ packages/shared-components/src/core/roving/index.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2026 Element Creations Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+export {
+    checkInputableElement,
+    findSiblingElement,
+    reducer,
+    RovingAction,
+    RovingTabIndexContext,
+    RovingTabIndexProvider,
+    Type,
+    useRovingTabIndex,
+} from "./RovingTabIndex";
+export type { IAction, IContext, IState, RovingTabIndexProviderProps } from "./RovingTabIndex";
+export { RovingTabIndexWrapper } from "./RovingTabIndexWrapper";
+export type { FocusHandler } from "./types";
